add disconnect telegram router

+++ app/api/telegram.py
@@ -1,7 +1,8 @@
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
+from app.core.jwt import create_access_token
 from app.models.telegram_account import TelegramAccount
-from app.schemas.telegram import PhoneAuthRequest, PhoneCodeVerifyRequest, TelegramConnect, TwoFactorAuthRequest
+from app.schemas.telegram import PhoneAuthRequest, PhoneCodeVerifyRequest,  TwoFactorAuthRequest
 from app.deps import get_current_user, get_db, get_telegram_service
 
 
@@ -9,7 +10,7 @@ router = APIRouter()
 
 
 @router.post("/auth/start")
-async def start_telegram_auth(request: PhoneAuthRequest, telegram_service=Depends(get_telegram_service)):
+async def start_telegram_auth(request: PhoneAuthRequest, user=Depends(get_current_user), telegram_service=Depends(get_telegram_service)):
     """
     Початок процесу авторизації - надсилання коду підтвердження
     """
@@ -30,8 +31,6 @@ async def verify_telegram_code(request: PhoneCodeVerifyRequest, db: Session = De
         request.phone_code_hash,
         request.session_string
     )
-
-    print("result", result)
 
     if result.get("status") == "success":
         account = TelegramAccount(
@@ -65,6 +64,9 @@ async def get_chats_telegram(db: Session = Depends(get_db), user=Depends(get_cur
     account = db.query(TelegramAccount).filter(
         TelegramAccount.user_id == user.id).first()
 
+    if not account or not account.is_telegram_auth:
+        return {"status": "telegram_not_autorize", "message": "Telegram account not authorized"}
+
     result = await telegram_service.get_chats(account.session_string)
 
     return result
@@ -75,20 +77,27 @@ async def get_messages_telegram(chat_id: int, db: Session = Depends(get_db), use
     account = db.query(TelegramAccount).filter(
         TelegramAccount.user_id == user.id).first()
 
-    print(chat_id)
+    if not account or not account.is_telegram_auth:
+        return {"status": "telegram_not_autorize", "message": "Telegram account not authorized"}
+
     result = await telegram_service.get_messages(chat_id, account.session_string)
 
     return result
 
 
-@router.post("/auth/disconnect")
+@router.get("/disconnect")
 async def disconnect_telegram(db: Session = Depends(get_db), user=Depends(get_current_user), telegram_service=Depends(get_telegram_service)):
     account = db.query(TelegramAccount).filter(
         TelegramAccount.user_id == user.id).first()
-    if not account:
-        raise HTTPException(
-            status_code=404, detail="Telegram account not found")
+
+    if not account or not account.is_telegram_auth:
+        return {"status": "telegram_not_autorize", "message": "Telegram account not authorized"}
+
+    telegram_service.logout(account.session_string)
+
     db.delete(account)
     db.commit()
 
-    return {"message": "Telegram account disconnected"}
+    access_token = create_access_token(data={"sub": user.email})
+
+    return {"isTelegramAuth": False, "token": access_token, "message": "Telegram account disconnected"}
